Added upgrade method to migrate from older versions. Plugin version is now stored as an option.

+++ admin/installer.php
@@ -13,10 +13,17 @@ class LSB_Installer {
         delete_option( WP_Options::STREAMS_BACKUP );
         delete_transient( WP_Options::STREAMS_TRANSIENT );
 
-        // cleanup legacy options <= 1.3
-        wp_clear_scheduled_hook( 'lsb_update_all_stream_status' );
-        delete_option('lsb-stream-storage');
-        delete_transient('lsb-stream-storage');
+        $previous_version = get_option( WP_Options::PLUGIN_VERSION ) ?: '1.3';
+        $this->upgrade( $previous_version, LSB_PLUGIN_VERSION );
+        update_option( WP_Options::PLUGIN_VERSION, LSB_PLUGIN_VERSION );
+    }
+    
+    function upgrade( $from, $to ) {
+        if ( $from === '1.3' ) {
+            wp_clear_scheduled_hook( 'lsb_update_all_stream_status' );
+            delete_option('lsb-stream-storage');
+            delete_transient('lsb-stream-storage');
+        }
     }
 
     function uninstall() {
